+++ src/main/java/christmas/domain/statistics/StatisticsGenerator.java
@@ -7,11 +7,7 @@ import christmas.domain.discount.DiscountCalculatorMapper;
 import christmas.domain.discount.DiscountResult;
 import christmas.domain.discount.DiscountResults;
 import christmas.domain.discount.DiscountType;
-import christmas.domain.gift.GiftCalculator;
-import christmas.domain.gift.GiftCalculatorMapper;
-import christmas.domain.gift.GiftResult;
 import christmas.domain.gift.GiftResults;
-import christmas.domain.gift.GiftType;
 import christmas.domain.menu.SelectedMenus;
 import christmas.domain.menu.SelectedMenusDTO;
 import java.time.LocalDate;
@@ -27,19 +23,10 @@ public final class StatisticsGenerator {
 
     public Statistics makeStatistics(LocalDate selectedDate, SelectedMenus selectedMenus) {
         DiscountResults discountResults = new DiscountResults(calculateDiscountResults(selectedDate, selectedMenus));
-        GiftResults giftResults = new GiftResults(calculateGiftResults(selectedMenus));
+        GiftResults giftResults = GiftResults.calculatedFrom(selectedMenus);
         Badge badge = badgeCalculator.calculate(discountResults, giftResults);
         SelectedMenusDTO selectedMenusDTO = selectedMenus.to();
         return new Statistics(selectedDate, selectedMenusDTO, giftResults, discountResults, badge);
-    }
-
-    private static List<GiftResult> calculateGiftResults(SelectedMenus selectedMenus) {
-        return Arrays.stream(GiftType.values()).map(giftType -> calculateGift(selectedMenus, giftType)).toList();
-    }
-
-    private static GiftResult calculateGift(SelectedMenus selectedMenus, GiftType giftType) {
-        GiftCalculator giftCalculator = GiftCalculatorMapper.map(giftType).orElseThrow();
-        return giftCalculator.calculate(selectedMenus);
     }
 
     private static List<DiscountResult> calculateDiscountResults(LocalDate selectedDate, SelectedMenus selectedMenus) {
